requirements.txt
# Research Agent System - Dependencies

# ===== LLM 프레임워크 =====
langgraph>=0.2.16
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0

# ===== LLM API =====
openai>=1.50.0
anthropic>=0.39.0
google-generativeai>=0.8.0

# ===== 데이터 분석 =====
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.5.0
statsmodels>=0.14.0
scipy>=1.13.0

# ===== 시각화 =====
matplotlib>=3.9.0
seaborn>=0.13.0
plotly>=5.24.0

# ===== 인용 관리 =====
bibtexparser>=2.0.0
requests>=2.32.0

# ===== PDF 생성 =====
reportlab>=4.2.0
markdown>=3.7.0

# ===== 유틸리티 =====
python-dotenv>=1.0.0
pyyaml>=6.0.0
aiohttp>=3.10.0
tqdm>=4.66.0

# ===== 웹 스크래핑 및 브라우저 자동화 =====
beautifulsoup4>=4.12.0
lxml>=5.3.0
playwright>=1.48.0
selenium>=4.26.0

# ===== 테스트 =====
pytest>=8.3.0
pytest-asyncio>=0.24.0
